+++ packages/react/src/createSeparateQuarkPropsFn.ts
@@ -1,11 +1,16 @@
-import { QuarkConfig } from '@quarkcss/core'
+import { getQuarkConfig, GetQuarkVariantsMap, QuarkCss } from '@quarkcss/core'
 
-export function createSeparateQuarkPropsFn<Config extends QuarkConfig>({ variants }: Config) {
+export function createSeparateQuarkPropsFn<Quark extends QuarkCss<{}>>(quark: Quark) {
   return <Props extends Record<string, any>>(
     props: Props
-  ): [Pick<Props, keyof Config['variants']>, Omit<Props, keyof Config['variants']>] => {
+  ): [
+    Pick<Props, keyof GetQuarkVariantsMap<Quark> & string>,
+    Omit<Props, keyof GetQuarkVariantsMap<Quark> & string>
+  ] => {
     const quarkProps = {} as any
     const rest = {} as any
+
+    const { variants } = getQuarkConfig(quark)
 
     if (!variants) return [quarkProps, props] as any